backend/requirements.txt
fastapi
uvicorn
google-generativeai
youtube-transcript-api
python-dotenv
pydantic